Add Connection Class
This adds logic to allow connecting two tickets together and verifying if they can be connected. The validation when adding will not be handled by this class, but instead a `ConnectedTicket` class.

If the asking ticket (from ticket) fails the lookup we increment the `bad_connects` value.

## Thoughts

Ideally `canConnect` should throw an error if the `bad_connects` are 10 or over. For now this can silently fail, but there should be an error page, or message that gets raised to the end user ideally. I'm not sure if there's a good way to go about that in this architecture.