Add ConnectedTicket Class
The `ConnectedTicket` class finds connections for the associated ticket that its instantiated with, as well as adds tickets.

Getting the `connections` will also include the joined ticket data via: `to_data`.

When adding a new connection if the ticket cannot connect to the given ticket the ticket's `bad_connects` is incremented by 1.